requirements.txt
groq==0.31.0
streamlit==1.32.1
python-dotenv==1.0.1
PyPDF2==3.0.1
pymupdf==1.24.0
pandas==2.2.1
numpy==1.26.4
